can't really hurt
svn path=/trunk/kdebase/kioslave/; revision=316998

+++ smb/kio_smb_browse.cpp
@@ -47,7 +47,7 @@ int SMBSlave::cache_stat(const SMBUrl &url, struct stat* st )
 {
     int result = smbc_stat( url.toSmbcUrl(), st);
     kdDebug(KIO_SMB) << "smbc_stat " << url.url() << " " << errno << " " << result << endl;
-    kdDebug(KIO_SMB) << "size " << st->st_size << endl;
+    kdDebug(KIO_SMB) << "size " << (long)st->st_size << endl;
     return result;
 }
